# src/main/resources/openapi/viewer-api.yaml
...
openapi: 3.0.3
info:
  title: 'Tailormap API'
  description: '
  Tailormap API contains the calls necessary to get the configuration for a map viewer and calls
  to get features.
  
  Note: you need to send an X-XSRF-TOKEN header with the value from the XSRF-TOKEN
  cookie for non-GET requests. The Angular HttpClient does this automatically.
  '
  contact:
    name: 'B3Partners BV'
    url: 'https://www.b3partners.nl/'
    email: 'info@b3partners.nl'
  license:
    name: 'MIT'
    url: 'https://mit-license.org/'
  version: 'v1'
servers:
  - url: '{basePath}'
    variables:
      basePath:
        description: base path
        default: '/api'
components:
  securitySchemes:
    formAuth:
      type: apiKey
      in: cookie
      name: JSESSIONID
  schemas:
...
    FeaturesResponse:
      description: 'A -possibly empty- list of features with requested attributes and presentation metadata'
      type: object
      required: [ features,
                  columnMetadata ]
      properties:
        total:
          description: '
            the total number of available features, can be an estimated value. 
            Can be missing when unknown or irrelevant (eg. featureinfo request)
          '
          type: integer
          nullable: true
        page:
          description: 'current page, this will allow the client to request the next or previous page by adding or removing 1 or more'
          type: integer
          nullable: true
          minimum: 1
        pageSize:
          description: 'number of features in a page, this will allow calculating the offset (page * pagesize)'
          type: integer
          nullable: true
          readOnly: true
          default: 100
          minimum: 1
        features:
          description: 'list of features. When the pagenumber*pagesize exceeds the number of features this list will be empty'
          type: array
          minItems: 0
          items:
            $ref: './viewer-schemas.yaml#/components/schemas/Feature'
        columnMetadata:
          description: 'presentation metadata'
          type: array
          minItems: 0
          items:
            $ref: './viewer-schemas.yaml#/components/schemas/ColumnMetadata'
        template:
          description: Template/description for a feature. Will replace placeholders by attribute values in the viewer
          type: string
          nullable: true
      example:
        total: 400
        page: 5
        pageSize: 100
        features: [ ]
        columnMetadata: [ ]
...
paths:
...
  /{viewerKind}/{name}/layer/{appLayerId}/features:
    summary: 'Use this endpoint to access features.'
    parameters:
      - in: path
        name: viewerKind
        required: true
        schema:
          type: string
          enum:
            - app
            - service
      - description: 'viewer name'
        in: path
        name: name
        required: true
        schema:
          type: string
      - in: path
        name: appLayerId
        required: true
        schema:
          type: string
      - description: 'x-coordinate, assumed in the coordinate reference system of the attribute source'
        in: query
        name: x
        required: false
        schema:
          type: number
      - description: 'y-coordinate, assumed in the coordinate reference system of the attribute source'
        in: query
        name: y
        required: false
        schema:
          type: number
      - description: 'buffer radius in the coordinate reference system of the attribute source'
        in: query
        name: distance
        required: false
        schema:
          type: number
          default: 4
      - description: 'feature id, the unique and immutable identifier of a feature within an attribute source'
        in: query
        name: __fid
        required: false
        schema:
          type: string
      - description: 'request true to have simplified geometries returned'
        in: query
        name: simplify
        required: false
        schema:
          type: boolean
          default: false
      - description: '
          A filter to be applied, possibly in combination with any other request parameters. 
          The filter is an ECQL string, see [ECQL reference](https://docs.geoserver.org/latest/en/user/filter/ecql_reference.html).
          Filtering is supported when requesting a page of features, not when requesting a single 
          feature (using `__fid`) nor when using x/y coordinates.
          '
        in: query
        name: filter
        required: false
        schema:
          type: string
      - description: '
          Page number to retrieve. For the initial request this should be 1 which will 
          then give the pageSize and total of features so additional pages can be retrieved.'
        in: query
        name: page
        required: false
        schema:
          type: integer
          minimum: 1
      - description: '
          The attribute name (not alias) to sort by. 
          The name must be of a configured attribute and not a geometry type.
          The default is to sort by the primary key of the feature.'
        in: query
        name: sortBy
        required: false
        schema:
          type: string
      - description: 'Sort order for sortBy'
        in: query
        name: sortOrder
        required: false
        schema:
          type: string
          default: asc
          enum:
            - asc
            - desc
      - description: Return only geometries, no attributes
        in: query
        name: onlyGeometries
        required: false
        schema:
          type: boolean
          default: false
      - description: Geometry output as part of the attributes, will be ignored if onlyGeometries is true
        in: query
        name: geometryInAttributes
        required: false
        schema:
          type: boolean
          default: false
    get:
      operationId: 'getFeatures'
      description: 'retrieve features that fulfill the requested conditions (parameters).'
      security:
        - formAuth: [ ]
      responses:
        '200':
          description: 'OK'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeaturesResponse'
        '400':
          description: 'Bad Request. May be returned for some combination of parameters that can not be processed or are incomplete.'
          content:
            application/json:
              schema:
                $ref: './status-responses.yaml#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: './status-responses.yaml#/components/schemas/RedirectResponse'
        '500':
          description: 'Internal server error'
          content:
            application/json:
              schema:
                $ref: './status-responses.yaml#/components/schemas/ErrorResponse'
    post:
      operationId: 'getFeaturesPOST'
      description: The same as GET, but allows for a large (spatial) filter.
      security:
        - formAuth: [ ]
      responses:
        '200':
          description: 'OK'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeaturesResponse'
        '400':
          description: 'Bad Request. May be returned for some combination of parameters that can not be processed or are incomplete.'
          content:
            application/json:
              schema:
                $ref: './status-responses.yaml#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: './status-responses.yaml#/components/schemas/RedirectResponse'
        '500':
          description: 'Internal server error'
          content:
            application/json:
              schema:
                $ref: './status-responses.yaml#/components/schemas/ErrorResponse'
...